Code/Projet/requirements.txt
numpy
matplotlib
scikit-learn
numpyencoder
Pillow
